Revert changes to the private implementation of entities.
